Cleaner updated to allow user list management.

+++ npvrd/src/pl/lodz/p/cm/ctp/npvrd/Cleaner.java
@@ -1,5 +1,8 @@
 package pl.lodz.p.cm.ctp.npvrd;
 
+import java.util.Calendar;
+import java.util.Date;
+import java.util.LinkedList;
 import java.util.List;
 import java.io.*;
 
@@ -13,6 +16,7 @@ public class Cleaner implements Runnable {
 	private TvChannelDAO tvChannelDAO;
 	private RecordingDAO recordingDAO;
 	private ProgramRecordingDAO programDvrScheduleDAO;
+	private UserRecordingDAO userRecordingDAO;
 	
 	public Cleaner() {
 		Npvrd.log(logPrefix + "Starting up");
@@ -20,6 +24,7 @@ public class Cleaner implements Runnable {
 		programDvrScheduleDAO = dbase.getProgramRecordingDAO();
 		tvChannelDAO = dbase.getTvChannelDAO();
 		recordingDAO = dbase.getRecordingDAO();
+		userRecordingDAO = dbase.getUserRecordingDAO();
 		try {
 			tvChannels = tvChannelDAO.list();
 		} catch (DAOException e) {
@@ -33,14 +38,48 @@ public class Cleaner implements Runnable {
 		for (TvChannel tc : tvChannels) {
 			try {
 				List<ProgramRecording> forDeletion = programDvrScheduleDAO.listOlderThanHours(tc.getId(), Npvrd.config.cleanerTolerance);
+				List<UserRecording> subscriptionsToLetGo = new LinkedList<UserRecording>();
+				int maxHold = Npvrd.config.cleanerMaxHold;
+				Date now = new Date();
+				
 				for (ProgramRecording cpr : forDeletion) {
-					File fileForDeletion = new File(Npvrd.config.recordings + cpr.recording.getFileName());
-					recordingDAO.delete(cpr.recording);
-					boolean result = fileForDeletion.delete();
-					if (!result) {
-						Npvrd.error(logPrefix + "Could not delete file " + cpr.recording.getFileName() + ". Dropped nonetheless.");
+					boolean deletable = true;
+					List<UserRecording> subscribers = userRecordingDAO.listByRecordingId(cpr.getRecording().getId());
+					for (UserRecording sub : subscribers) {
+						Calendar cal = Calendar.getInstance();
+						if (sub.getCreatedOn().before(cpr.getProgram().getEnd())) {
+							//Timestamp EndPlusHold = (Timestamp)cpr.getProgram().getEnd().clone();
+							Date endPlusHold = new Date(cpr.getProgram().getEnd().getTime());
+							cal.setTime(endPlusHold);
+							cal.add(Calendar.HOUR_OF_DAY, maxHold);
+							if (cal.after(now)) {
+								deletable = false;
+							} else {
+								subscriptionsToLetGo.add(sub);
+							}
+						} else {
+							Date createdOnPlusHold = (Date)sub.getCreatedOn().clone();
+							cal.setTime(createdOnPlusHold);
+							cal.add(Calendar.HOUR_OF_DAY, maxHold);
+							if (cal.after(now)) {
+								deletable = false;
+							} else {
+								subscriptionsToLetGo.add(sub);
+							}
+						}
 					}
-					counter++;
+					for (UserRecording sub : subscriptionsToLetGo) {
+						userRecordingDAO.delete(sub);
+					}
+					if (deletable) {
+						File fileForDeletion = new File(Npvrd.config.recordings + cpr.getRecording().getFileName());
+						recordingDAO.delete(cpr.getRecording());
+						boolean result = fileForDeletion.delete();
+						if (!result) {
+							Npvrd.error(logPrefix + "Could not delete file " + cpr.getRecording().getFileName() + ". Dropped nonetheless.");
+						}
+						counter++;
+					}
 				}
 			} catch (DAOException e) {
 				Npvrd.error(logPrefix + "Problem working with the database in Cleaner");
